Combine test_charms and test_snaps into test_cdk
